refactor(config): move CandidateLoader to config and remove nullable chaining
